backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
python-dotenv==1.0.1
Werkzeug==3.0.4
passlib[bcrypt]==1.7.4
pgvector==0.3.5

